Fixes #52: Report dates did not update after internal navigation
When the report model hydrated, the date did not change leading to all report dates looking the same. On hard refresh, the dates would update because the data model was updated. The date was translated into a timestamp, which wasn't being reactively updated. Denoting it as as derrived fixes the issue.